Move `CpbReaderV2` and `CpbLoaderV2` under package that is the only one to make use of them
